Moves the skills field to the Skills Section component
